Initial check-in for the synchronization of subprogram specifications M604-054
This is the initial check-in for the feature and it contains two major
features:

* Modular extension mechanism for editor buffers:

It is now possible to extend the functionality of editor buffers in GPS
core, by the way of Editor_Listeners. One can create a new Listener
object that will react to specific events sent by each instance of
editor buffers.

This mechanism is not specific to the GPS GUI, and is part of GPS core,
so it will be possible :

- To use listeners made in Ada in GnatBench (the sync spec listener for
  example)
- To create listeners in Java provided the interface is exposed.

* Synchronization of subprogram specifications between body and
  definition.

This feature is deactivated for now as it is not yet usable. It uses the
previously mentioned listener mechanism, but the listener isn't yet
registered.

The way the feature works as of now is very basic : If the user is in a
spec and modifies it, the listener will impact the modifications on its
counterpart, in the definition/body, using heuristics to preserve the
existing layout of the counterpart.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@213204 936e1b1b-40f2-da11-902a-00137254ae57