Add support for checking the xref database schema version number
We use sqlite's standard pragma user_version to store the version, and
compare it every time gnatinspect is spawned. When the version does not
match, an error message is displayed. The user can use the new switch
--check_db_version to automatically delete obsolete databases and
recreate them (this is the default in GPS).
One update to the schema, by adding a new 'project' field in the
'files' table, in preparation for support of aggregate projects in
gnatinspect (MC19-005).

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@222798 936e1b1b-40f2-da11-902a-00137254ae57