("="): fix handling of aggregate projects
Comparing the project nodes only works if they are in the same tree, which
is not true for homonym projects in different aggregated trees.
This fixes Include_File, which would not insert a file multiple times
when it has the same full path in multiple projects.

(Create): now sets ambiguous to true when the same full path is found in
 multiple aggregate projects. We still return the common full path, but
 a tool like gnatinspect needs to know that the object files (and thus
 ALI files) will be ambiguous in this case.

This will also be tested as part of the test for MC19-005

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@223433 936e1b1b-40f2-da11-902a-00137254ae57