(Info_Set): never return an empty set
The previous hange (returning an empty set when the file is not a project
source) broke FB03_003, which expects at least some guesses regarding the
language of the file.

("GPS.execute_action"): change the error message to be the filter's error
message, when available, instead of a generic 'invalid context for action'.
This provides more information in general.

N220-036

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224012 936e1b1b-40f2-da11-902a-00137254ae57