Add proper support for aggregate projects
Each editor view is now associated with a specific project, so that tooltips
 and xref can be properly displayed. We can thus have multiple view of the
 same file each in its own project.
When using aggregates, we display the name of the project in the MDI tab.
Creating a new view of an editor uses the same project as the initial view
 (including when creating from dnd)
omni-search displays name of the project (when using aggregates) for filenames
 and sources contexts. Duplicate files will appear several times in the list
 of results, once per project.
Tooltips now use the view's project, instead of the first possible project.
 This is more accurate.
Hyper-links also use the view's context.
Project information for the view is saved in the perspective.
A source editor always sets the project information in the context (it did not
 before). To preserve the same contextual menus, however, it sets a flag so
 that Has_Project_Information returns False in this case.

N305-035

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224357 936e1b1b-40f2-da11-902a-00137254ae57