(gnatinspect, Get_Entity): file name can now be omitted to search in all files
It is possible to search for the entity in all the source files of the
project by passing a empty string for the name of the file. Searching
for predefined entities now requires a specific value for File (although
the previous use of the empty string would still work in most cases).

N409-022

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@225944 936e1b1b-40f2-da11-902a-00137254ae57