Rewrite gnatls handling
Most of this handling was specific to GPS, so did not apply to gnatinspect.
Instead, the code has now been moved to GNATCOLL.Projects, and will benefit
all users of that API.
The code will now properly spawn gnatls when it detects that the project
specifies an alternate command line, which is the case for cross-environment
and in particular when using a different runtime (--RTS). In this case,
GNATCOLL spawns gnatls, updates its project path, and parses the project
again.
N507-006

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@227440 936e1b1b-40f2-da11-902a-00137254ae57