Rewrite the test L406_006
This test checks cross-references with runtime files, but is in fact not
reliable on our build machines since the order in the output depends on
the exact full path of the test directory and runtime directory.
We therefore simulate a local predefined directory to achieve the same
effect.
Looking at the full path of the files that contain the body of Proc, it
was in fact discovered that parsing runtime ALI files would still point
to the local source file that overrides them (always utils.adb instead
of runtime/utils.adb when parsing runtime/utils.ali for instance), so
this was also fixed in the source code of gnatcoll-xref.adb and
gnatcoll-projects.adb.
N610-001

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@228843 936e1b1b-40f2-da11-902a-00137254ae57