Use the new Runtime and Target attributes to compute the gnatls to spawn
When any of these two attribute is defined, GNATCOLL will spawn
 <target>-gnatls --RTS=<runtime>
(omitting any of the component when the variable is not defined).

It is an error for a project file to defined any of these two variable
in addition to the Ide'Gnatlist attribute.

Users can override the value of Runtime and Target via the new
Set_Target_And_Runtime procedure on the environment. This is correctly
taken into account by Get_Runtime and Get_Target.

N918-040

Change-Id: If66bc44ebeaf1a39a4f4f6c449b995e8e76ce0a5