Review implementation of smart pointers
A new package Shared_Pointers is added, more flexible than the old
Smart_Pointers package. The Element_Type no longer needs to be a tagged
type derived from Refcounted. By using a local storage pool, we are able
to limit the number of memory allocations, and this new implementation
is faster than the old one, while more flexible:  40% of the time it
used to take to create 300_000 smart pointers, 40% of the time for
the assignment, and about the same time for Get.

A new storage_pool was also added, which can allocate memory for the
element and a special header for each chunk of memory. This header can
be used to keep information like the refcount, or the next and previous
pointers for a list.

The new smart_pointers are now used everywhere, as a replacement for the
shared_pointers. In particular, this fixes one case where we were doing
the refcounting manually, and not using atomic operations for that
purpose.

No longer rename gnatcoll-atomic__*.ada to gnatcoll-atomic.adb. Instead,
we use a proper naming scheme directly in the project. This is more flexible
when we need to edit projects.

O422-013

Change-Id: Ie8449dd972c4d6f6ffc0d06a13d10a546de42a60