Do not define the Separate_Suffix attribute
There used to be a bug there (we were specifing an "Ada" index,
which in fact the project manager is rejecting). But removing that
indexes means that all .adb files are always part of an Ada project,
even if the user's naming scheme changes the body suffix.
In fact, gprconfig's knowledge base never defines that attribute either,
so it is safe to remove it.

Should fix various regressions introduced by M304-034

Change-Id: I8e4d4c5dff694e6fcf0492cc5f441d9a05dc918e