Force static libiconv when using --disable-shared and --with-iconv=path
When both switches are specified, the tools like gnatinspect will be
built with gnatcoll static libraries. They should also using a static
libiconv, to avoid conflicts when running on other systems.

This patch has no impact when using --with-iconv=path and not
--disable-shared, since in this case we want the tools to be linked
against shared libraries, including for iconv.

This patch also has no effect when using --disable-shared and --with-iconv
(i.e. using auto-search), to keep it simpler and avoid breaking existing
setups. The recommended solution, in any case, is to use --with-iconv=path
as has now been made clear in the doc.

This patch supercedes using --with-iconv=static, which did not provide
a way to specify an explicit directory and only supported auto-search.

O923-012

Change-Id: I9b8398cd4e7958fbd796a0ee17e54f10809a64b0