Do not emit block prefix if appending an empty string
The block prefix is emitted when the length of the current block is
zero. So, don't emit it when appending a 0-length substring (which
may happen when Passthrough is called if the first character of the
input string needs encoding), otherwise we will emit it again for
the next (non-empty) substring.

Adjustment on previous change for OC14-010

Change-Id: I55168b2268280d507b89de6dd7bc2ddc7a098e18