P608-011 return empty target when it is not specified
Target attribute is always inherited by project file from cgpr.
So if Attribute_Value for "target" is not empty we need
to additionally check whether it comes from given project
(or ones extended by it) or from cgpr.
Add Default_To_Host flag to Get_Target, if set to True
and target not specified - return host. If set to False
and target not specified - return empty string.

Change-Id: Ie27e7e9480d3ecf11bf5df42de0d38594f007ad2