P629-052 (1/2) introduce IDE_Mode switch
It appears that sometimes IDEs and single pass applications
have a bit different needs when loading the project.
For example, when loading two projects with same scenario
variable that has different set of possible values we do not
want to see warnings about them in GPS (see IC01_004 and
IC01_004_2) so a filter was added under P421_026. But we do
want those warnings in i.e. gnatcheck if a wrong value was
passed via -X switch.
A least disruptive solution is to introduce an IDE_Mode switch
that should be specified during project environment
initialization.
Switch is off by default and it disables the warning filtering
mentioned above. All other upcoming cases when we will need a
slightly special behaviour by GPS but not by other tools will
also fall under this switch.
Comments state explicitly that the switch is only of interest
to programs like GPS.

Change-Id: I75f8d4d246bc96b1837e114b59c5ea10bde213da