P905-013 adapt project load to changes in gprconfig
Now that bareboard targets do not have a default runtime
specifying only the target but not runtime leads to the same
thing as to not specifying target at all and not having
native compiler on path.

Update projects:Test:test_bareboard test.out accordingly.

Add call to GPR.Conf.Set_Runtime_For in Recompute_View to get rid
of junk error message about "compiler not found for default runtime"
when in fact runtime was explicitly set.


Change-Id: I07ec382de62f316b8f7902b9e897ed7b02abb92d