Add support for casting values in CREATE TABLE AS
Postgres, in particular, needs to be able to infer column types, but
when we use
   CREATE TEMP TABLE t AS VALUES (1, 'foo'), (2, 'bar');
it cannot always do so (in the above example it would work, but if for
instance we represent a UUID, which looks like a string, the temporary
table would have a column of type text and not uuid.

Instead, we now generate
   CREATE TEMP TABLE t AS VALUES (1::integer, 'foo'::text), (2, 'bar');
in postgres, or
   CREATE TEMP TABLE t AS VALUES (cast (1 as integer), cast ('foo' as text)),
      (2, 'bar');

Moved some tests from testsuite/sql/queries to testsuite/sql/sql1, so
that we can test the different output for sqlite and postgres. This also
highlights differences in the output of the money type.