Change the use of colors in gnatcoll.traces
Colors are no longer specified as a string, but as a (Fg, Bg, Style)
 record.
This avoids duplication with GNATCOLL.Terminal, for the ANSI sequences.
This also means that we can more easily store colors. For instance, a
 block_trace_handle now displays the "entering" and "leaving" messages
 in dim, to make them less obtrusive.
This will also let us configure per-handle colors in the future.

(Assert): clarify documentation
 This procedure should not be used just to get red output !

This change is slightly backward incompatible:
 (Trace) that takes a string for the color still exists (and works as
    before).
 (Increase_Indent, Decrease_Indent): removed versions that take colors
    as strings.
 (Create block handler): also removed versions that take colors as
    strings.